requirements.txt
# requirements.txt
markdown
python-dotenv
streamlit
openai>=1.0
